test: add TUI theme transparency unit tests and refactor logic
